user cfn't edit other user's post and photo

+++ app/Http/Controllers/Dashboard/PostController.php
@@ -53,22 +53,20 @@ class PostController extends Controller
      */
     public function store(ValidatePostFormRequest $request)
     {
-        if (Auth::user()->hasRole('user')) {
-            $post = Post::create([
-                'title'         => $request->title,
-                'description'   => $request->description,
-                'visability'    => $request->visability,
-                'user_id'       => Auth::user()->id
-            ]);
+        $post = Post::create([
+            'title'         => $request->title,
+            'description'   => $request->description,
+            'visability'    => $request->visability,
+            'user_id'       => Auth::user()->id
+        ]);
 
-            if($post) {
-                return redirect('/dashboard/posts')
-                    ->with('success', 'New post added successfully');
+        if($post) {
+            return redirect('/dashboard/posts')
+                ->with('success', 'New post added successfully');
 
-            } else {
-                return redirect('/dashboard/posts')
-                ->with('error', 'Something went wrong');
-            }
+        } else {
+            return redirect('/dashboard/posts')
+            ->with('error', 'Something went wrong');
         }
     }
 
@@ -80,11 +78,16 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        if ($post) {
-            return view('dashboard.posts.show', compact('post'));
+        if (Auth::user()->id == $post->user_id) {
+            if ($post) {
+                return view('dashboard.posts.show', compact('post'));
+            } else {
+                return redirect('/dashboard/posts')
+                    ->with('error', 'Post not found');
+            }
         } else {
             return redirect('/dashboard/posts')
-                ->with('error', 'Post not found');
+                ->with('error', 'You do not have necessary permissions');
         }
     }
 
@@ -96,7 +99,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        return view('dashboard.posts.edit', compact('post'));
+        if (Auth::user()->id == $post->user_id) {
+            return view('dashboard.posts.edit', compact('post'));
+        } else {
+            return redirect('/dashboard/posts')
+            ->with('error', 'You do not have necessary permissions');
+        }    
     }
 
     /**
@@ -108,25 +116,30 @@ class PostController extends Controller
      */
     public function update(ValidatePostFormRequest $request, Post $post)
     {
-        if ($post) {
-            $post->title         =  $request->title;
-            $post->description   =  $request->description;
-            $post->visability    =  $request->visability;
-            $post->save();
-
+        if (Auth::user()->id == $post->user_id) {
             if ($post) {
-                return redirect('dashboard/posts')
-                    ->with('success', 'Post updated successfully');
+                $post->title         =  $request->title;
+                $post->description   =  $request->description;
+                $post->visability    =  $request->visability;
+                $post->save();
+
+                if ($post) {
+                    return redirect('dashboard/posts')
+                        ->with('success', 'Post updated successfully');
+
+                } else {
+                    return redirect('/dashboard/posts')
+                        ->with('error', 'Something went wrong');
+                }
 
             } else {
                 return redirect('/dashboard/posts')
-                    ->with('error', 'Something went wrong');
+                    ->with('error', 'Post not found');
             }
-
         } else {
             return redirect('/dashboard/posts')
-                ->with('error', 'Post not found');
-        }
+            ->with('error', 'You do not have necessary permissions');
+        }  
     }
 
     /**
@@ -137,34 +150,47 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        if ($post) {
-            $post->delete();
+        if (Auth::user()->id == $post->user_id || Auth::user()->hasRole('admin')) {
+            if ($post) {
+                $post->delete();
 
-            return redirect('dashboard/posts')
-                ->with('success', 'Post deleted successfully');
+                return redirect('dashboard/posts')
+                    ->with('success', 'Post deleted successfully');
 
+            } else {
+                return redirect('/dashboard/posts')
+                    ->with('error', 'Something went wrong');
+            } 
         } else {
             return redirect('/dashboard/posts')
-                ->with('error', 'Something went wrong');
-        }
+            ->with('error', 'You do not have necessary permissions');
+        }  
     }
 
-    public function updateVisability(Request $request)
+    public function updateVisability(Request $request, Post $post)
     {
-        if ($request->mode == 'true') {
-            DB::table('posts')->where('id', $request->id)
-                ->update([
-                    'visability' => '1'
-                ]);
-        }
-        else {
-            DB::table('posts')->where('id', $request->id)
-                ->update(['visability' => '0']);
-        }
+        $post     = Post::where('id', $request->id)->first();
+        $user     = User::where('id', $post->user_id)->first();
 
-        return response()->json([
-            'message'  => 'Post visability was successfully updated',
-            'status'   => 200
-        ]);
+        if (Auth::user()->id == $user->id || Auth::user()->hasRole('admin')) {
+            if ($request->mode == 'true') {
+                DB::table('posts')->where('id', $request->id)
+                    ->update([
+                        'visability' => '1'
+                    ]);
+            }
+            else {
+                DB::table('posts')->where('id', $request->id)
+                    ->update(['visability' => '0']);
+            }
+
+            return response()->json([
+                'message'  => 'Post visability was successfully updated',
+                'status'   => 200
+            ]);
+        } else {
+            return redirect('/dashboard/posts')
+            ->with('error', 'You do not have necessary permissions');
+        }  
     }
 }
